search logic

+++ ios/Runner/Encryption/EncryptionService.swift
@@ -10,7 +10,7 @@ import Foundation
 import RxSwift
 
 protocol EncryptionService {
-    func generateKeyPairAndStore(expireAt: Double) -> Observable<Bool>
+    func generateKeyPairAndStore(expireAt: Int64) -> Observable<Bool>
     func encrypt(publicKeyStr: String, message: String) -> Observable<String>
     func decrypt(ciphertext: String) -> Observable<String>
     var publicKey: String? {get}
